+++ packages/dialect-react-ui/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - fix: consistent `useDappAudience` return type
 
+
+## [1.0.0-beta.35] - 2022-08-09
+- fix: notifications header
+
 ## [1.0.0-beta.34] - 2022-08-09
 
 - fix: tune a bit wallet settings field